feat: init refacto
