+++ carbonserver/carbonserver/api/infra/database/sql_models.py
@@ -129,7 +129,6 @@ class Organization(Base):
     id = Column(UUID(as_uuid=True), primary_key=True, index=True, default=uuid.uuid4)
     name = Column(String)
     description = Column(String)
-    api_key = Column(String)
     projects = relationship("Project", back_populates="organization")
 
     def __repr__(self):
@@ -145,8 +144,6 @@ class User(Base):
     id = Column(UUID(as_uuid=True), primary_key=True, index=True, default=uuid.uuid4)
     name = Column(String)
     email = Column(String, unique=True, index=True)
-    hashed_password = Column(String)
-    api_key = Column(String)
     is_active = Column(Boolean, default=True)
     organizations = Column(ARRAY(String, as_tuple=False, dimensions=1))
