+++ carbonserver/carbonserver/api/services/user_service.py
@@ -1,14 +1,14 @@
 from typing import List
 
 from carbonserver.api.infra.repositories.repository_users import SqlAlchemyRepository
-from carbonserver.api.schemas import User, UserAuthenticate, UserCreate
+from carbonserver.api.schemas import User, UserAutoCreate
 
 
 class UserService:
     def __init__(self, user_repository: SqlAlchemyRepository) -> None:
         self._repository: SqlAlchemyRepository = user_repository
 
-    def create_user(self, user: UserCreate) -> User:
+    def create_user(self, user: UserAutoCreate) -> User:
         created_user: User = self._repository.create_user(user)
 
         return created_user
@@ -21,6 +21,3 @@ class UserService:
         users_list = self._repository.list_users()
 
         return users_list
-
-    def verify_user(self, user: UserAuthenticate) -> bool:
-        return self._repository.verify_user(user)
